+++ src/queries/jangter/index.ts
@@ -7,7 +7,6 @@ import {
 import { useNavigate } from 'react-router-dom';
 
 import {
-  addJangterBookmarks,
   createProduct,
   deleteJangterBookmarks,
   deleteProduct,
@@ -186,32 +185,6 @@ export const useUpdateteProductStatus = (productId: number) => {
       // 요청 성공 시 실행할 로직
       queryClient.invalidateQueries({
         queryKey: ['products', productId, 'detail'],
-      });
-    },
-    onError: (error) => {
-      // 요청 실패 시 실행할 로직
-      console.error('Mutation failed:', error);
-    },
-    onSettled: () => {
-      // 요청 완료 후 (성공/실패 관계없이) 실행할 로직
-    },
-  });
-};
-
-export const useAddBookmark = (productId: number) => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: async () => {
-      await addJangterBookmarks(productId);
-    },
-    onSuccess: () => {
-      // 요청 성공 시 실행할 로직
-      console.log('Mutation succeeded:', productId);
-      queryClient.invalidateQueries({
-        predicate: (query) =>
-          Array.isArray(query.queryKey) &&
-          query.queryKey[0] === 'jangterBookmarks',
       });
     },
     onError: (error) => {
